app/requirements.txt
amqp==5.1.1
ansi2html==1.8.0
appnope==0.1.3
APScheduler==3.10.1
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==4.1.0
cachelib==0.9.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
colorcet==3.0.1
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
dash==2.11.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-extensions==1.0.1
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-table==5.0.0
dask==2023.6.1
datashader==0.15.0
datashape==0.5.2
debugpy==1.6.7
decorator==5.1.1
EasyProcess==1.1
EditorConfig==0.12.3
entrypoint2==1.1
exceptiongroup==1.1.2
executing==1.2.0
Flask==2.2.5
Flask-Caching==2.0.2
fonttools==4.40.0
fsspec==2023.6.0
geographiclib==2.0
geojson==2.5.0
geopy==2.3.0
gunicorn==20.0.4
hatanaka==2.8.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.23.3
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsbeautifier==1.14.8
jupyter-dash==0.4.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
kombu==5.3.1
llvmlite==0.40.1
locket==1.0.0
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
more-itertools==9.1.0
multipledispatch==1.0.0
ncompress==1.0.1
nest-asyncio==1.5.6
numba==0.57.1
numpy==1.24.0
obspy==1.4.0
orjson==3.9.1
packaging==23.1
pandas==1.5.3
param==1.13.0
parso==0.8.3
partd==1.4.0
patool==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
plotly==5.15.0
plotly-resampler==0.8.3.2
pluggy==1.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyct==0.5.0
Pygments==2.15.1
pyowm==3.3.0
pyparsing==3.1.0
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyunpack==0.3
PyYAML==6.0
pyzmq==25.1.0
redis==4.6.0
requests==2.31.0
retrying==1.3.4
scipy==1.11.1
six==1.16.0
SQLAlchemy==2.0.17
stack-data==0.6.2
tenacity==8.2.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
trace-updater==0.0.9.1
traitlets==5.9.0
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==5.0.1
unlzw3==0.2.2
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wget==3.2
xarray==2023.6.0
zipp==3.15.0
